fix: resolve comment without handle logic offer with different credentials
