+++ src/core/agent/agent.types.ts
@@ -152,6 +152,7 @@ interface KeriaNotification {
   multisigId?: string;
   connectionId: string;
   read: boolean;
+  linkedGroupRequests?: Record<string, boolean>;
 }
 
 enum KeriConnectionType {
